[refactor] components 아래 폴더 생성, 파일 이동
