+++ Frontend/src/components/MeetingRoom/Memo.jsx
