+++ Frontend/src/components/MeetingRoom/StreamBox.jsx
