+++ Frontend/src/components/MeetingRoom/StreamBox.module.css
